+++ src/vivarium_testing_utils/automated_validation/interface.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 import pandas as pd
 
-from vivarium_testing_utils.automated_validation import plot_utils
+from vivarium_testing_utils.automated_validation.visualization import plot_utils
 from vivarium_testing_utils.automated_validation.comparison import Comparison, FuzzyComparison
 from vivarium_testing_utils.automated_validation.data_loader import DataLoader, DataSource
 from vivarium_testing_utils.automated_validation.data_transformation.calculations import (
