+++ tests/automated_validation/visualization/test_plot_utils.py
@@ -4,7 +4,10 @@ import pandas as pd
 import matplotlib.pyplot as plt
 from matplotlib.testing.compare import compare_images
 import os
-from vivarium_testing_utils.automated_validation.plot_utils import line_plot, RELPLOT_KWARGS
+from vivarium_testing_utils.automated_validation.visualization.plot_utils import (
+    line_plot,
+    RELPLOT_KWARGS,
+)
 import seaborn as sns
 from vivarium_testing_utils.automated_validation.data_loader import DataSource
